GameItem, GameInventory components and refactoring
