+++ components/GameGear.vue
@@ -1,0 +1,11 @@
+<template>
+  <div></div>
+</template>
+
+<script lang="ts" setup>
+import { usePlayerStore } from "@/stores/player";
+
+const playerStore = usePlayerStore();
+</script>
+
+<style lang="scss" scoped></style>
