+++ components/GameInventory.vue
@@ -1,7 +1,32 @@
 <template>
-  <div></div>
+  <div class="inventory-container">
+    <div
+      v-for="(inventoryItem, index) in playerStore.inventory"
+      :key="index"
+      class="inv-slot"
+    >
+      <GameItem
+        v-if="inventoryItem && inventoryItem?.itemId !== null"
+        :inventory-item="inventoryItem!"
+      />
+      <GameItem v-else :empty-slot="true" />
+    </div>
+  </div>
 </template>
 
-<script lang="ts" setup></script>
+<script lang="ts" setup>
+import { usePlayerStore } from "@/stores/player";
 
-<style lang="scss" scoped></style>
+const playerStore = usePlayerStore();
+</script>
+
+<style lang="scss" scoped>
+.inventory-container {
+  height: fit-content;
+  width: fit-content;
+  display: grid;
+  grid-template-columns: 1fr 1fr 1fr 1fr;
+  grid-template-rows: 1fr 1fr 1fr 1fr 1fr 1fr 1fr;
+  gap: 10px;
+}
+</style>
